requirements.txt
discord.py
pytz
aiosqlite